providing the signin button on the product listing, if i'm signed out
